# backend/src/main/resources/openapi/swagger.yaml
openapi: 3.0.3
info:
  title: Steam Game Catalog API
  description: API de consultation du catalogue de jeux Steam
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
tags:
  - name: Games
    description: Consultation et suppression des jeux
  - name: Balance
    description: Gestion du solde utilisateur
  - name: Features
    description: Mise en avant du catalogue
  - name: Orders
    description: Gestion des commandes
  - name: Admin - Users
    description: Gestion des utilisateurs (proxié vers le service d'authentification Laravel). Requiert un token Bearer admin.
  - name: Admin - Orders
    description: Gestion des commandes (CRUD direct sur game_db). Requiert un token Bearer admin.
  - name: Admin - Games
    description: Gestion du catalogue de jeux (CRUD direct sur game_db). Requiert un token Bearer admin.
paths:
  /games:
    get:
      tags:
        - Games
      summary: Récupérer le catalogue de jeux (GetAll)
      description: Liste paginée des jeux avec filtres optionnels.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          description: Recherche full-text
          schema:
            type: string
        - name: genreId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Liste paginée récupérée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageGameSummary'

  /games/{id}:
    get:
      tags:
        - Games
      summary: Obtenir TOUS les détails d'un jeu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Détails complets du jeu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameDetail'
        '404':
          description: Jeu non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Games
      summary: Supprimer un jeu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Jeu supprimé avec succès
        '404':
          description: Jeu non trouvé

  /games/{id}/key_counts:
    get:
      tags:
        - Games
      summary: Obtenir le nombre de clés disponibles pour un jeu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  key_counts:
                    type: integer

  /balance:
    get:
      tags:
        - Balance
      summary: Obtenir le solde
      parameters:
        - name: userid
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Solde actuel
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: integer
    post:
      tags:
        - Balance
      summary: Modifier le solde
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userid:
                  type: integer
                amount:
                  type: integer
      responses:
        '200':
          description: Résultat de l'opération
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /feature:
    get:
      tags:
        - Features
      summary: Obtenir les sélections de jeux de la page d'accueil (Paginé)
      description: Retourne les différentes listes de jeux (meilleures ventes, nouveautés, promotions, à venir) sous forme paginée.
      parameters:
        - name: page
          in: query
          description: Numéro de la page (0-indexed)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Nombre d'éléments par page pour chaque catégorie
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureData'

  /orders:
    get:
      tags:
        - Orders
      summary: Obtenir toutes les commandes d'un utilisateur
      parameters:
        - name: userid
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste des commandes
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
    post:
      tags:
        - Orders
      summary: Créer une commande
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userid:
                  type: integer
                games:
                  type: array
                  items:
                    type: object
                    properties:
                      game_ids:
                        type: integer
                      quantity:
                        type: integer
      responses:
        '200':
          description: Résultat de la commande
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string

  /games/genres:
    get:
      tags:
        - Games
      summary: Obtenir la liste de tous les genres
      responses:
        '200':
          description: Liste des genres
          content:
            application/json:
              schema:
                type: object
                properties:
                  genres:
                    type: array
                    items:
                      $ref: '#/components/schemas/Genre'

  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Obtenir une commande spécifique
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userid
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détail de la commande
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'

  /admin/users:
    get:
      tags:
        - Admin - Users
      summary: Lister tous les utilisateurs
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des utilisateurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUser'
        '403':
          description: Accès interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
        - Admin - Users
      summary: Créer un utilisateur
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserRequest'
      responses:
        '201':
          description: Utilisateur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '403':
          description: Accès interdit

  /admin/users/{id}:
    get:
      tags:
        - Admin - Users
      summary: Obtenir un utilisateur par ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Utilisateur trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '403':
          description: Accès interdit
        '404':
          description: Utilisateur non trouvé
    put:
      tags:
        - Admin - Users
      summary: Mettre à jour un utilisateur
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserRequest'
      responses:
        '200':
          description: Utilisateur mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '403':
          description: Accès interdit
        '404':
          description: Utilisateur non trouvé
    delete:
      tags:
        - Admin - Users
      summary: Supprimer un utilisateur
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Utilisateur supprimé
        '403':
          description: Accès interdit
        '404':
          description: Utilisateur non trouvé

  /admin/orders:
    get:
      tags:
        - Admin - Orders
      summary: Lister toutes les commandes
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des commandes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminOrderDto'
        '403':
          description: Accès interdit
    post:
      tags:
        - Admin - Orders
      summary: Créer une commande
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminOrderRequest'
      responses:
        '201':
          description: Commande créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderDto'
        '403':
          description: Accès interdit
        '404':
          description: Jeu ou utilisateur non trouvé

  /admin/orders/{id}:
    get:
      tags:
        - Admin - Orders
      summary: Obtenir une commande par ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Commande trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderDto'
        '403':
          description: Accès interdit
        '404':
          description: Commande non trouvée
    put:
      tags:
        - Admin - Orders
      summary: Mettre à jour une commande
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminOrderRequest'
      responses:
        '200':
          description: Commande mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderDto'
        '403':
          description: Accès interdit
        '404':
          description: Commande non trouvée
    delete:
      tags:
        - Admin - Orders
      summary: Supprimer une commande
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Commande supprimée
        '403':
          description: Accès interdit
        '404':
          description: Commande non trouvée

  /admin/games:
    get:
      tags:
        - Admin - Games
      summary: Lister tous les jeux
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des jeux
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameSummary'
        '403':
          description: Accès interdit
    post:
      tags:
        - Admin - Games
      summary: Créer un jeu
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminGameRequest'
      responses:
        '201':
          description: Jeu créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSummary'
        '400':
          description: Données invalides
        '403':
          description: Accès interdit

  /admin/games/{id}:
    get:
      tags:
        - Admin - Games
      summary: Obtenir les détails complets d'un jeu (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Détails du jeu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGameDetail'
        '403':
          description: Accès interdit
        '404':
          description: Jeu non trouvé
    put:
      tags:
        - Admin - Games
      summary: Mettre à jour un jeu
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminGameRequest'
      responses:
        '200':
          description: Jeu mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGameDetail'
        '400':
          description: Données invalides
        '403':
          description: Accès interdit
        '404':
          description: Jeu non trouvé
    delete:
      tags:
        - Admin - Games
      summary: Supprimer un jeu (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Jeu supprimé
        '403':
          description: Accès interdit
        '404':
          description: Jeu non trouvé

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Token Passport OAuth2 émis par le service d'authentification Laravel. Le rôle "admin" est requis.
  schemas:
    GameSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        steamAppId:
          type: integer
        name:
          type: string
        slug:
          type: string
        headerImage:
          type: string
        priceFinal:
          type: integer
        reductionPercentage:
          type: integer
        recommendationsTotal:
          type: integer
        releaseDate:
          type: string
          format: date
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'

    GameDetail:
      allOf:
        - $ref: '#/components/schemas/GameSummary'
        - type: object
          description: Détails complets d'un jeu, incluant toutes les relations SQL.
          properties:
            detailedDescription:
              type: string
              description: Description HTML complète
            aboutTheGame:
              type: string
              description: Description HTML courte
            supportedLanguages:
              type: string
            requiredAge:
              type: integer
            releaseDate:
              type: string
              format: date
              nullable: true
            releaseDateRaw:
              type: string
              description: Date de sortie sous forme de texte brut (ex "TBD", "Q1 2024")
            metacriticScore:
              type: integer
              nullable: true
            recommendationsTotal:
              type: integer
            currency:
              type: string
              example: "EUR"
            priceInitial:
              type: integer
              description: Prix de base avant réduction, en centimes.
            pcRequirements:
              type: object
              additionalProperties: true
              description: Configuration requise PC (format JSON)
            macRequirements:
              type: object
              additionalProperties: true
            linuxRequirements:
              type: object
              additionalProperties: true
            updatedAt:
              type: string
              format: date-time
              description: Date de dernière mise à jour des données
            companies:
              type: array
              items:
                $ref: '#/components/schemas/GameCompany'
            categories:
              type: array
              items:
                $ref: '#/components/schemas/Category'
            screenshots:
              type: array
              items:
                $ref: '#/components/schemas/Screenshot'
            movies:
              type: array
              items:
                $ref: '#/components/schemas/Movie'
    FeatureData:
      type: object
      description: Objet contenant les différentes catégories de jeux mises en avant, toutes paginées.
      properties:
        topseller:
          $ref: '#/components/schemas/PageGameSummary'
        new_release:
          $ref: '#/components/schemas/PageGameSummary'
        low_deals:
          type: object
          description: Jeux en promotion classés par tranches de prix (paginés)
          properties:
            under_5:
              $ref: '#/components/schemas/PageGameSummary'
            under_10:
              $ref: '#/components/schemas/PageGameSummary'
            under_20:
              $ref: '#/components/schemas/PageGameSummary'
        upcoming:
          $ref: '#/components/schemas/PageGameSummary'

    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        gamesId:
          type: integer
          format: int64
        key:
          type: string
        game:
          $ref: '#/components/schemas/GameInfo'

    GameInfo:
      type: object
      properties:
        name:
          type: string
        headerImage:
          type: string
        genres:
          type: array
          items:
            type: string

    GameCompany:
      type: object
      properties:
        company:
          $ref: '#/components/schemas/Company'
        role:
          type: string

    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

    Genre:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string

    Category:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string

    PageGameSummary:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/GameSummary'
        totalElements:
          type: integer
        totalPages:
          type: integer

    ApiError:
      type: object
      properties:
        status:
          type: integer
          description: Code HTTP de l'erreur
        error:
          type: string
          description: Code d'erreur (ex. NOT_FOUND, BAD_REQUEST)
        message:
          type: string
          description: Message d'erreur détaillé
        path:
          type: string
          description: Chemin de la requête qui a provoqué l'erreur
        timestamp:
          type: string
          format: date-time
          description: Timestamp de l'erreur

    Screenshot:
      type: object
      required:
        - pathThumbnail
        - pathFull
      properties:
        id:
          type: integer
          format: int64
          description: BIGSERIAL PRIMARY KEY
        steamId:
          type: integer
          description: INTEGER (steam_id)
        pathThumbnail:
          type: string
          format: uri
          description: TEXT NOT NULL (path_thumbnail)
        pathFull:
          type: string
          format: uri
          description: TEXT NOT NULL (path_full)
        position:
          type: integer
          description: SMALLINT (position)

    Movie:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: BIGSERIAL PRIMARY KEY
        steamId:
          type: integer
          description: INTEGER (steam_id)
        name:
          type: string
          description: VARCHAR(500) (name)
        thumbnail:
          type: string
          format: uri
          description: TEXT (thumbnail)
        dashAv1:
          type: string
          format: uri
          description: TEXT (dash_av1)
        dashH264:
          type: string
          format: uri
          description: TEXT (dash_h264)
        hlsH264:
          type: string
          format: uri
          description: TEXT (hls_h264)
        highlight:
          type: boolean
          description: BOOLEAN (highlight)
        position:
          type: integer
          description: SMALLINT (position)

    AdminUser:
      type: object
      description: Utilisateur géré par le service d'authentification Laravel.
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]

    AdminUserRequest:
      type: object
      description: Payload pour créer ou mettre à jour un utilisateur.
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        role:
          type: string
          enum: [user, admin]
          default: user

    AdminOrderDto:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        gamesId:
          type: integer
          format: int64
        key:
          type: string

    AdminOrderRequest:
      type: object
      required:
        - userId
        - gamesId
        - key
      properties:
        userId:
          type: integer
        gamesId:
          type: integer
          format: int64
        key:
          type: string

    AdminGameRequest:
      type: object
      description: Payload pour créer ou mettre à jour un jeu.
      properties:
        steamAppId:
          type: integer
        name:
          type: string
        slug:
          type: string
        headerImage:
          type: string
          format: uri
        priceFinal:
          type: integer
          description: Prix en centimes
        reductionPercentage:
          type: integer
        releaseDate:
          type: string
          format: date
          nullable: true
        releaseDateRaw:
          type: string
          description: Texte brut de la date (ex. "TBD", "Q1 2024")
        requiredAge:
          type: integer
        metacriticScore:
          type: integer
          nullable: true
        recommendationsTotal:
          type: integer
        currency:
          type: string
          example: EUR
        platformWindows:
          type: boolean
        platformMac:
          type: boolean
        platformLinux:
          type: boolean
        detailedDescription:
          type: string
        aboutTheGame:
          type: string
        supportedLanguages:
          type: string
        pcRequirements:
          type: string
          description: JSON sérialisé de la configuration requise PC
        macRequirements:
          type: string
        linuxRequirements:
          type: string

    AdminGameDetail:
      type: object
      description: Détails complets d'un jeu exposés à l'admin (inclut les champs de plateforme).
      properties:
        id:
          type: integer
          format: int64
        steamAppId:
          type: integer
        name:
          type: string
        slug:
          type: string
        headerImage:
          type: string
          format: uri
        priceFinal:
          type: integer
        reductionPercentage:
          type: integer
        recommendationsTotal:
          type: integer
        releaseDate:
          type: string
          format: date
          nullable: true
        releaseDateRaw:
          type: string
        requiredAge:
          type: integer
        metacriticScore:
          type: integer
          nullable: true
        currency:
          type: string
        platformWindows:
          type: boolean
        platformMac:
          type: boolean
        platformLinux:
          type: boolean
        detailedDescription:
          type: string
        aboutTheGame:
          type: string
        supportedLanguages:
          type: string
        pcRequirements:
          type: object
          additionalProperties: true
        macRequirements:
          type: object
          additionalProperties: true
        linuxRequirements:
          type: object
          additionalProperties: true